Rename rechannel to renet
Remove renet, replaced by rechannel.
