+++ bevy_renet/Cargo.toml
@@ -12,7 +12,7 @@ edition = "2021"
 
 [dependencies]
 bevy = { version = "0.9.0", default-features = false }
-renet = { path = "../renet", version = "0.0.10", features = ["bevy"] }
+renet = { path = "../renet", version = "0.0.10" }
 
 [dev-dependencies]
 serde = { version = "1.0", features = [ "derive" ] }
