--- rechannel/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "rechannel"
-version = "0.0.6"
-authors = ["Lucas Poffo <ltkpoffo@gmail.com>"]
-description = "Server/Client network library for reliable communication channels in multiplayer games"
-keywords = ["gamedev", "networking"]
-repository = "https://github.com/lucaspoffo/renet"
-readme = "README.md"
-license = "MIT OR Apache-2.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-log = "0.4.17"
-bytes = "1.1"
-octets = "0.2"
-renetcode = { path = "../renetcode", version = "0.0.6" }
-
-[dev-dependencies]
-env_logger = "0.9.0"
-rand = "0.8.5"
