--- rechannel/src/lib.rs
@@ -1,10 +1,0 @@
-pub mod channels;
-mod connection_stats;
-pub mod error;
-mod packet;
-pub mod remote_connection;
-pub mod server;
-
-pub mod transport;
-
-pub use bytes::Bytes;
