+++ renet/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "renet"
 version = "0.0.10"
+authors = ["Lucas Poffo <ltkpoffo@gmail.com>"]
 description = "Server/Client network library for multiplayer games with authentication and connection management"
 keywords = ["gamedev", "networking"]
 repository = "https://github.com/lucaspoffo/renet"
@@ -8,15 +9,13 @@ readme = "../README.md"
 license = "MIT OR Apache-2.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[features]
-bevy = ["dep:bevy_ecs"]
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
-rechannel = { path = "../rechannel", version = "0.0.6" }
-renetcode = { path = "../renetcode", version = "0.0.6" }
 log = "0.4.17"
-bevy_ecs = { version = "0.9.0", optional = true }
+bytes = "1.1"
+octets = "0.2"
+renetcode = { path = "../renetcode", version = "0.0.6" }
 
 [dev-dependencies]
 env_logger = "0.9.0"
